[IMP] im_livechat: remove AbstractMessage (merge with WebsiteLivechatMessage)
Also rename WebsiteLivechatMessage to PublicLivechatMessage

Task-2894499
